+++ addon/components/fd-attribute-table.js
@@ -1,5 +1,6 @@
 import Component from '@ember/component';
 import { A } from '@ember/array';
+import { isNone } from '@ember/utils';
 import layout from '../templates/components/fd-attribute-table';
 
 export default Component.extend({
@@ -63,7 +64,10 @@ export default Component.extend({
     */
     deleteValue() {
       let selectedValues = this.get('selectedValues');
-      this.get('delete')(selectedValues);
+      let deleteAction = this.get('delete');
+      if (!isNone(deleteAction)) {
+        deleteAction(selectedValues);
+      }
 
       selectedValues.forEach((value) => value.deleteRecord());
       selectedValues.clear();