+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 ### Added
 * Collapse button for class primitives.
 
+### Changed
+* The diagram in the `{{fd-uml-diagram}}` component now has an infinite size.
+* Links can't have empty source or target.
+
 ## [0.7.0-beta.1] - 2019-05-17
 ### Added
 * Added json update when editing diagrams.